CVS_SILENT fix compiler warning (as frerich would say, this is surely worth a new release ;)
svn path=/trunk/koffice/; revision=222761

+++ lib/kotext/kotextobject.cc
@@ -1071,7 +1071,7 @@ KCommand * KoTextObject::setAlignCommand( KoTextCursor * cursor, int align, int 
         return 0L;
     KoTextDocument * textdoc = textDocument();
     if ( !textdoc->hasSelection( selectionId, true ) && cursor &&
-         cursor->parag()->alignment() == align )
+         (int)cursor->parag()->alignment() == align )
         return 0L; // No change needed.
 
     emit hideCursor();
